requirements.txt
openai>=1.40.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
